feat(testing-environment): extract applyProp from preact-upstream-tests
Move the string-key → Element PAPI dispatch logic (applyViaElementPAPI)
out of preact-upstream-tests/setup-nocompile.js and into
@lynx-js/testing-environment as a public export `applyProp()`, making it
reusable for future framework adapters.
